requirements.txt
asgiref==3.6.0
Django==4.2
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.4
